server/requirement.txt
langchain
langchain-openai
langchain-core
langchain-community
streamlit
python-dotenv
langchain-ollama
langserve
fastapi
uvicorn
sse_starlette
pydantic
bs4
pypdf 
chromadb
faiss-cpu
langchain-openai
arxiv
wikipedia
groq
langchain-groq
beautifulsoup4
sentence_transformers
langchain-huggingface
unstructured
libmagic